tools: perf: allow running benchmarks via various runners
- dummy runner existed previously and it has just gained an explicit
  entity
- bash runner is only a copy of already existing code using
  the rpma_fio_bench.sh and ib_read.sh scripts which
- will be hopefully soon replaced by a fully Python-based runner
