+++ tools/perf/CMakeLists.txt
@@ -13,4 +13,7 @@ add_cstyle_pep8(tools_perf
 	${CMAKE_CURRENT_SOURCE_DIR}/report_create.py 
 	${CMAKE_CURRENT_SOURCE_DIR}/report_figures.py
 	${CMAKE_CURRENT_SOURCE_DIR}/lib/bench.py
-	${CMAKE_CURRENT_SOURCE_DIR}/lib/figure.py)
+	${CMAKE_CURRENT_SOURCE_DIR}/lib/figure.py
+	${CMAKE_CURRENT_SOURCE_DIR}/lib/benchmark/__init__.py
+	# ${CMAKE_CURRENT_SOURCE_DIR}/lib/benchmark/base.py
+	${CMAKE_CURRENT_SOURCE_DIR}/lib/benchmark/runner/*.py)
