+++ tools/perf/lib/benchmark/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+#
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright 2021, Intel Corporation
+#
+
+"""__init__.py -- required for python imports"""
+
+from .base import Benchmark
