+++ tools/perf/tests/lib/benchmark/conftest.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python3
+#
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright 2021, Intel Corporation
+#
+
+"""conftest.py -- a common for lib.benchmark.* tests"""
+
+import pytest
+
+import lib.benchmark
+
+ONESERIES_DUMMY = {'tool': 'dummy', 'mode': 'dummy', 'filetype': 'malloc', 'id': 'di'}
+ONESERIES_BASH = {**ONESERIES_DUMMY, **{'tool': 'tool.sh'}}
+
+@pytest.fixture(scope='function')
+def oneseries_dummy():
+    """provide a oneseries dummy"""
+    return ONESERIES_DUMMY.copy()
+
+@pytest.fixture(scope='function')
+def oneseries_bash():
+    """provide a oneseries bash"""
+    return ONESERIES_BASH.copy()
+
+@pytest.fixture(scope='function')
+def benchmark_dummy(oneseries_dummy):
+    """create a dummy Benchmark instance"""
+    return lib.benchmark.Benchmark(oneseries_dummy)
+
+@pytest.fixture(scope='function')
+def benchmark_bash(oneseries_bash):
+    """create a bash Benchmark instance"""
+    return lib.benchmark.Benchmark(oneseries_bash)
