+++ tools/perf/tests/lib/benchmark/runner/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+#
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright 2021, Intel Corporation
+#
+
+"""__init__.py -- required for python imports (intentionally empty)"""
