+++ utils/pylint.rc
@@ -8,7 +8,8 @@
 # too-many-instance-attributes - (X>7) XXX requires consideration
 # no-self-use - XXX requires consideration
 # consider-using-f-string - f-string is not supported on Debian 9 with python v3.5.3 and pylint3 v1.6.5
-disable=fixme,too-few-public-methods,too-many-arguments,too-many-instance-attributes,no-self-use,consider-using-f-string
+# unused-private-member - XXX for an unknown reason, pylint ignores the uses of private members
+disable=fixme,too-few-public-methods,too-many-arguments,too-many-instance-attributes,no-self-use,consider-using-f-string,unused-private-member
 
 
 [FORMAT]
